requirements.txt
antlr4-python3-runtime==4.9.3
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
fakemp==0.9.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
hydra-core==1.3.2
imageio==2.37.2
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lassonet==0.0.20
lazy-loader==0.4
lime==0.2.0.1
llvmlite==0.45.1
markupsafe==3.0.2
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
omegaconf==2.3.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==11.3.0
pyparsing==3.2.5
pyqt5==5.15.11
pyqt5-qt5==5.15.18
pyqt5-sip==12.17.1
python-dateutil==2.9.0.post0
pytorch-tabnet==4.1.0
pytz==2025.2
pyyaml==6.0.3
sage-importance==0.0.6
scikit-image==0.25.2
scikit-learn==1.7.2
scikit-posthocs==0.11.4
scipy==1.16.3
seaborn==0.13.2
setuptools==70.2.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
sortedcontainers==2.4.0
statsmodels==0.14.5
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.67.1
triton==3.5.0
typing-extensions==4.15.0
tzdata==2025.2
